initial work on game logic
Finally got the code of the game logic working. Basically, I used four
functions to check the text to find a match then prints in the console
a response. If they don’t match, the console produces a different
response. Took way to long to figure out a simple solution. I can now
work on scoring and we’ll have the first functional version of TT!